afl: Ensure running scripts from aflplusplus/

+++ aflplusplus/scripts/_cmin.sh
@@ -4,6 +4,10 @@
 # then feed it back to the fuzzers.
 # rsync is used instead of cp
 # because cp doesn't like overwriting files.
+
+DESIRED_PATH="aflplusplus"
+[[ "${PWD##*/}" == "$DESIRED_PATH" ]] || { echo "Run from $DESIRED_PATH"; exit; }
+
 mkdir -p queue_all
 rsync --progress -a --delete findings/fuzzer*/queue/* queue_all/ || exit
 afl-cmin -i queue_all -o queue_cmin -- build/aflplusplus \
